+++ EmmaSharp/Methods/Members.cs
@@ -99,13 +99,13 @@ namespace EmmaSharp
         /// <param name="memberId">Member identifier.</param>
 		/// <returns>Member opt out date and mailing if member is opted out.</returns>
 		/// <remarks>Http404 if no member is found.</remarks>
-		public MemberOptout GetMemberOptout(string memberId)
+		public List<MemberOptout> GetMemberOptout(string memberId)
 		{
 			var request = new RestRequest();
 			request.Resource = "/{accountId}/members/{memberId}/optout";
 			request.AddUrlSegment("memberId", memberId);
 
-			return Execute<MemberOptout>(request);
+			return Execute<List<MemberOptout>>(request);
 		}
 
         /// <summary>
@@ -267,16 +267,18 @@ namespace EmmaSharp
         /// Add a single member to one or more groups.
         /// </summary>
         /// <param name="memberId">Member identifier.</param>
-        /// <param name="members">Group ids to which to add this member.</param>
+        /// <param name="groupIds">Group ids to which to add this member.</param>
         /// <returns>An array of ids of the affected groups.</returns>
         /// <remarks>Http404 if no member is found.</remarks>
-        public List<int> AddMemberToGroups(string memberId, MemberGroups members)
+        public List<int> AddMemberToGroups(string memberId, List<int> groupIds)
         {
             var request = new RestRequest(Method.PUT);
             request.Resource = "/{accountId}/members/{memberId}/groups";
             request.AddUrlSegment("memberId", memberId);
 
-            request.AddBody(members);
+            request.RequestFormat = DataFormat.Json;
+            request.JsonSerializer = new EmmaJsonSerializer();
+            request.AddBody(new { group_ids = groupIds });
 
             return Execute<List<int>>(request);
         }
@@ -285,10 +287,10 @@ namespace EmmaSharp
         /// Remove a single member from one or more groups.
         /// </summary>
         /// <param name="memberId">Member identifier.</param>
-        /// <param name="members">Group ids from which to remove this member</param>
+        /// <param name="groupIds">Group ids from which to remove this member</param>
         /// <returns>An array of references to the affected groups.</returns>
         /// <remarks>Http404 if no member is found.</remarks>
-        public List<int> RemoveMemberFromGroups(string memberId, MemberGroups members)
+        public List<int> RemoveMemberFromGroups(string memberId, List<int> groupIds)
         {
             var request = new RestRequest(Method.PUT);
             request.Resource = "/{accountId}/members/{memberId}/groups/remove";
@@ -296,7 +298,7 @@ namespace EmmaSharp
 
             request.RequestFormat = DataFormat.Json;
             request.JsonSerializer = new EmmaJsonSerializer();
-            request.AddBody(members);
+            request.AddBody(new { group_ids = groupIds });
 
             return Execute<List<int>>(request);
         }
@@ -498,7 +500,11 @@ namespace EmmaSharp
             request.AddUrlSegment("statusTo", statusTo.ToEnumString<MemberStatusShort>());
 
             if (!string.IsNullOrWhiteSpace(groupId))
-                request.AddParameter("group_id", groupId);
+            {
+                request.RequestFormat = DataFormat.Json;
+                request.JsonSerializer = new EmmaJsonSerializer();
+                request.AddBody(new { group_id = groupId });
+            }
 
             return Execute<bool>(request);
         }