+++ EmmaSharp/Models/Members/Member.cs
@@ -16,10 +16,10 @@ namespace EmmaSharp.Models.Members
         public DateTime? ConfirmedOptIn { get; set; }
 
         [JsonProperty("account_id")]
-        public int? AccountId { get; set; }
+        public long? AccountId { get; set; }
 
         [JsonProperty("fields")]
-        public Dictionary<string, string> Fields { get; set; }
+        public Dictionary<string, object> Fields { get; set; }
 
         [JsonProperty("member_id")]
         public long? MemberId { get; set; }